Remove warnings

+++ lib/GameArea.h
@@ -31,7 +31,6 @@ class GameArea : public gf::Transformable {
 		gf::Texture textureBackground;
         gf::Sprite background;
 
-
 	public :
 
 		void updateTextureBackground(Grid & gd);
@@ -53,17 +52,11 @@ class GameArea : public gf::Transformable {
 	        for (size_t i = 0; i < WIDTH_G; ++i){
 	        	for (size_t j = 0; j < HEIGHT_G; ++j){
 	                gf::Sprite sprite;
-	                sprite.setPosition({i*SIZE_CASE, j*SIZE_CASE});
+	                sprite.setPosition({(float)i*SIZE_CASE, (float)j*SIZE_CASE});
 	                (*this)(i, j) = sprite;
 	            }
 	        }
-
-
 		};
-
-
-
-
 };
